fix: valid destinations not displayed in some cases
If you click on a piece, the valid destinations are colored. If you click on an empty cell the colored cells go back to their default color.
If then you click again on the same piece as before, the valid destination cells are not colored because the "selected piece" was not reset when clicking on the empty cell, and the game is set to do nothing if you click on the same piece multiple times.